fix: add balance refresh on menu screen, reuse refresh logic (#366)
