+++ packages/cryptoplease/lib/features/profile/menu_screen.dart
@@ -1,3 +1,4 @@
+import 'package:cryptoplease/app/components/refresh_balance_wrapper.dart';
 import 'package:cryptoplease/features/profile/components/menu_header.dart';
 import 'package:cryptoplease/features/profile/components/menu_section.dart';
 import 'package:cryptoplease_ui/cryptoplease_ui.dart';
@@ -7,19 +8,21 @@ class MenuScreen extends StatelessWidget {
   const MenuScreen({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context) => CpTheme.light(
-        child: ColoredBox(
-          color: Colors.white,
-          child: SafeArea(
-            bottom: false,
-            child: ColoredBox(
-              color: CpColors.lightGreyBackground,
-              child: ListView(
-                physics: const ClampingScrollPhysics(),
-                children: const [
-                  MenuHeader(),
-                  MenuSection(),
-                ],
+  Widget build(BuildContext context) => RefreshBalancesWrapper(
+        builder: (context, _) => CpTheme.light(
+          child: ColoredBox(
+            color: Colors.white,
+            child: SafeArea(
+              bottom: false,
+              child: ColoredBox(
+                color: CpColors.lightGreyBackground,
+                child: ListView(
+                  physics: const ClampingScrollPhysics(),
+                  children: const [
+                    MenuHeader(),
+                    MenuSection(),
+                  ],
+                ),
               ),
             ),
           ),
